+++ CHANGES.rst
@@ -5,6 +5,8 @@ Changelog
 1.2.52 (unreleased)
 -------------------
 
+- Update to Plone 6.1.5 and buidlout 5-x versions
+
 - WEBBDC-2831 : Update some fields descriptions
   [boulch]